Backend/requirement.txt
fastapi
uvicorn
pydantic
redis
python-dotenv
pandas
prophet